LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/animations/background-position-origin-interpolation-expected.txt


FAIL CSS Transitions: property <background-position> from neutral to [left 20px top 20px] at (0) should be [10px 10px] assert_equals: expected "10px 10px " but got "left 10px top 10px "
FAIL CSS Transitions: property <background-position> from neutral to [left 20px top 20px] at (0.25) should be [12.5px 12.5px] assert_equals: expected "12.5px 12.5px " but got "left 12.5px top 12.5px "
FAIL CSS Transitions: property <background-position> from neutral to [left 20px top 20px] at (0.5) should be [15px 15px] assert_equals: expected "15px 15px " but got "left 15px top 15px "
FAIL CSS Transitions: property <background-position> from neutral to [left 20px top 20px] at (0.75) should be [17.5px 17.5px] assert_equals: expected "17.5px 17.5px " but got "left 17.5px top 17.5px "
FAIL CSS Transitions: property <background-position> from neutral to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
FAIL CSS Transitions with transition: all: property <background-position> from neutral to [left 20px top 20px] at (0) should be [10px 10px] assert_equals: expected "10px 10px " but got "left 10px top 10px "
FAIL CSS Transitions with transition: all: property <background-position> from neutral to [left 20px top 20px] at (0.25) should be [12.5px 12.5px] assert_equals: expected "12.5px 12.5px " but got "left 12.5px top 12.5px "
FAIL CSS Transitions with transition: all: property <background-position> from neutral to [left 20px top 20px] at (0.5) should be [15px 15px] assert_equals: expected "15px 15px " but got "left 15px top 15px "
FAIL CSS Transitions with transition: all: property <background-position> from neutral to [left 20px top 20px] at (0.75) should be [17.5px 17.5px] assert_equals: expected "17.5px 17.5px " but got "left 17.5px top 17.5px "
FAIL CSS Transitions with transition: all: property <background-position> from neutral to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
PASS CSS Animations: property <background-position> from neutral to [left 20px top 20px] at (0) should be [10px 10px] 
PASS CSS Animations: property <background-position> from neutral to [left 20px top 20px] at (0.25) should be [12.5px 12.5px] 
PASS CSS Animations: property <background-position> from neutral to [left 20px top 20px] at (0.5) should be [15px 15px] 
PASS CSS Animations: property <background-position> from neutral to [left 20px top 20px] at (0.75) should be [17.5px 17.5px] 
PASS CSS Animations: property <background-position> from neutral to [left 20px top 20px] at (1) should be [20px 20px] 
PASS Web Animations: property <background-position> from neutral to [left 20px top 20px] at (0) should be [10px 10px] 
PASS Web Animations: property <background-position> from neutral to [left 20px top 20px] at (0.25) should be [12.5px 12.5px] 
FAIL Web Animations: property <background-position> from neutral to [left 20px top 20px] at (0.5) should be [15px 15px] assert_equals: expected "15px 15px " but got "16.25px 16.25px "
FAIL Web Animations: property <background-position> from neutral to [left 20px top 20px] at (0.75) should be [17.5px 17.5px] assert_equals: expected "17.5px 17.5px " but got "19.06px 19.06px "
PASS Web Animations: property <background-position> from neutral to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL CSS Transitions: property <background-position> from [initial] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "left 0px top 0px "
FAIL CSS Transitions: property <background-position> from [initial] to [left 20px top 20px] at (0.25) should be [5px 5px] assert_equals: expected "5px 5px " but got "left 5px top 5px "
FAIL CSS Transitions: property <background-position> from [initial] to [left 20px top 20px] at (0.5) should be [10px 10px] assert_equals: expected "10px 10px " but got "left 10px top 10px "
FAIL CSS Transitions: property <background-position> from [initial] to [left 20px top 20px] at (0.75) should be [15px 15px] assert_equals: expected "15px 15px " but got "left 15px top 15px "
FAIL CSS Transitions: property <background-position> from [initial] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
FAIL CSS Transitions with transition: all: property <background-position> from [initial] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "left 0px top 0px "
FAIL CSS Transitions with transition: all: property <background-position> from [initial] to [left 20px top 20px] at (0.25) should be [5px 5px] assert_equals: expected "5px 5px " but got "left 5px top 5px "
FAIL CSS Transitions with transition: all: property <background-position> from [initial] to [left 20px top 20px] at (0.5) should be [10px 10px] assert_equals: expected "10px 10px " but got "left 10px top 10px "
FAIL CSS Transitions with transition: all: property <background-position> from [initial] to [left 20px top 20px] at (0.75) should be [15px 15px] assert_equals: expected "15px 15px " but got "left 15px top 15px "
FAIL CSS Transitions with transition: all: property <background-position> from [initial] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
FAIL CSS Animations: property <background-position> from [initial] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "0px 0px "
PASS CSS Animations: property <background-position> from [initial] to [left 20px top 20px] at (0.25) should be [5px 5px] 
PASS CSS Animations: property <background-position> from [initial] to [left 20px top 20px] at (0.5) should be [10px 10px] 
PASS CSS Animations: property <background-position> from [initial] to [left 20px top 20px] at (0.75) should be [15px 15px] 
PASS CSS Animations: property <background-position> from [initial] to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL Web Animations: property <background-position> from [initial] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "0px 0px "
PASS Web Animations: property <background-position> from [initial] to [left 20px top 20px] at (0.25) should be [5px 5px] 
PASS Web Animations: property <background-position> from [initial] to [left 20px top 20px] at (0.5) should be [10px 10px] 
PASS Web Animations: property <background-position> from [initial] to [left 20px top 20px] at (0.75) should be [15px 15px] 
PASS Web Animations: property <background-position> from [initial] to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL CSS Transitions: property <background-position> from [inherit] to [left 20px top 20px] at (0) should be [80px 80px] assert_equals: expected "80px 80px " but got "left 80px top 80px "
FAIL CSS Transitions: property <background-position> from [inherit] to [left 20px top 20px] at (0.25) should be [65px 65px] assert_equals: expected "65px 65px " but got "left 65px top 65px "
FAIL CSS Transitions: property <background-position> from [inherit] to [left 20px top 20px] at (0.5) should be [50px 50px] assert_equals: expected "50px 50px " but got "left 50px top 50px "
FAIL CSS Transitions: property <background-position> from [inherit] to [left 20px top 20px] at (0.75) should be [35px 35px] assert_equals: expected "35px 35px " but got "left 35px top 35px "
FAIL CSS Transitions: property <background-position> from [inherit] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
FAIL CSS Transitions with transition: all: property <background-position> from [inherit] to [left 20px top 20px] at (0) should be [80px 80px] assert_equals: expected "80px 80px " but got "left 80px top 80px "
FAIL CSS Transitions with transition: all: property <background-position> from [inherit] to [left 20px top 20px] at (0.25) should be [65px 65px] assert_equals: expected "65px 65px " but got "left 65px top 65px "
FAIL CSS Transitions with transition: all: property <background-position> from [inherit] to [left 20px top 20px] at (0.5) should be [50px 50px] assert_equals: expected "50px 50px " but got "left 50px top 50px "
FAIL CSS Transitions with transition: all: property <background-position> from [inherit] to [left 20px top 20px] at (0.75) should be [35px 35px] assert_equals: expected "35px 35px " but got "left 35px top 35px "
FAIL CSS Transitions with transition: all: property <background-position> from [inherit] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
PASS CSS Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0) should be [80px 80px] 
PASS CSS Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0.25) should be [65px 65px] 
PASS CSS Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0.5) should be [50px 50px] 
PASS CSS Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0.75) should be [35px 35px] 
PASS CSS Animations: property <background-position> from [inherit] to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL Web Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0) should be [80px 80px] assert_equals: expected "80px 80px " but got "10px 10px "
FAIL Web Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0.25) should be [65px 65px] assert_equals: expected "65px 65px " but got "12.5px 12.5px "
FAIL Web Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0.5) should be [50px 50px] assert_equals: expected "50px 50px " but got "16.25px 16.25px "
FAIL Web Animations: property <background-position> from [inherit] to [left 20px top 20px] at (0.75) should be [35px 35px] assert_equals: expected "35px 35px " but got "19.06px 19.06px "
PASS Web Animations: property <background-position> from [inherit] to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL CSS Transitions: property <background-position> from [unset] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "left 0px top 0px "
FAIL CSS Transitions: property <background-position> from [unset] to [left 20px top 20px] at (0.25) should be [5px 5px] assert_equals: expected "5px 5px " but got "left 5px top 5px "
FAIL CSS Transitions: property <background-position> from [unset] to [left 20px top 20px] at (0.5) should be [10px 10px] assert_equals: expected "10px 10px " but got "left 10px top 10px "
FAIL CSS Transitions: property <background-position> from [unset] to [left 20px top 20px] at (0.75) should be [15px 15px] assert_equals: expected "15px 15px " but got "left 15px top 15px "
FAIL CSS Transitions: property <background-position> from [unset] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
FAIL CSS Transitions with transition: all: property <background-position> from [unset] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "left 0px top 0px "
FAIL CSS Transitions with transition: all: property <background-position> from [unset] to [left 20px top 20px] at (0.25) should be [5px 5px] assert_equals: expected "5px 5px " but got "left 5px top 5px "
FAIL CSS Transitions with transition: all: property <background-position> from [unset] to [left 20px top 20px] at (0.5) should be [10px 10px] assert_equals: expected "10px 10px " but got "left 10px top 10px "
FAIL CSS Transitions with transition: all: property <background-position> from [unset] to [left 20px top 20px] at (0.75) should be [15px 15px] assert_equals: expected "15px 15px " but got "left 15px top 15px "
FAIL CSS Transitions with transition: all: property <background-position> from [unset] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
FAIL CSS Animations: property <background-position> from [unset] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "0px 0px "
PASS CSS Animations: property <background-position> from [unset] to [left 20px top 20px] at (0.25) should be [5px 5px] 
PASS CSS Animations: property <background-position> from [unset] to [left 20px top 20px] at (0.5) should be [10px 10px] 
PASS CSS Animations: property <background-position> from [unset] to [left 20px top 20px] at (0.75) should be [15px 15px] 
PASS CSS Animations: property <background-position> from [unset] to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL Web Animations: property <background-position> from [unset] to [left 20px top 20px] at (0) should be [0% 0%] assert_equals: expected "0 % 0 % " but got "0px 0px "
PASS Web Animations: property <background-position> from [unset] to [left 20px top 20px] at (0.25) should be [5px 5px] 
PASS Web Animations: property <background-position> from [unset] to [left 20px top 20px] at (0.5) should be [10px 10px] 
PASS Web Animations: property <background-position> from [unset] to [left 20px top 20px] at (0.75) should be [15px 15px] 
PASS Web Animations: property <background-position> from [unset] to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px top 20px] at (0.25) should be [calc(5px + 37.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 37.5 % + 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px top 20px] at (0.5) should be [calc(10px + 25%) calc(10px + 25%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 25 % + 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px top 20px] at (0.75) should be [calc(15px + 12.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 12.5 % + 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px top 20px] at (0.25) should be [calc(5px + 37.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 37.5 % + 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px top 20px] at (0.5) should be [calc(10px + 25%) calc(10px + 25%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 25 % + 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px top 20px] at (0.75) should be [calc(15px + 12.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 12.5 % + 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px top 20px] at (1) should be [20px 20px] assert_equals: expected "20px 20px " but got "left 20px top 20px "
PASS CSS Animations: property <background-position> from [center center] to [left 20px top 20px] at (0) should be [50% 50%] 
FAIL CSS Animations: property <background-position> from [center center] to [left 20px top 20px] at (0.25) should be [calc(5px + 37.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 37.5 % + 5px ) " but got "calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px top 20px] at (0.5) should be [calc(10px + 25%) calc(10px + 25%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 25 % + 10px ) " but got "calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px top 20px] at (0.75) should be [calc(15px + 12.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 12.5 % + 15px ) " but got "calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
PASS CSS Animations: property <background-position> from [center center] to [left 20px top 20px] at (1) should be [20px 20px] 
PASS Web Animations: property <background-position> from [center center] to [left 20px top 20px] at (0) should be [50% 50%] 
FAIL Web Animations: property <background-position> from [center center] to [left 20px top 20px] at (0.25) should be [calc(5px + 37.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 37.5 % + 5px ) " but got "calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [left 20px top 20px] at (0.5) should be [calc(10px + 25%) calc(10px + 25%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 25 % + 10px ) " but got "calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [left 20px top 20px] at (0.75) should be [calc(15px + 12.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 12.5 % + 15px ) " but got "calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
PASS Web Animations: property <background-position> from [center center] to [left 20px top 20px] at (1) should be [20px 20px] 
FAIL CSS Transitions: property <background-position> from [center center] to [center top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [center top 20px] at (0.25) should be [50% calc(5px + 37.5%)] assert_equals: expected "50 % calc ( 37.5 % + 5px ) " but got "50 % top calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [center top 20px] at (0.5) should be [50% calc(10px + 25%)] assert_equals: expected "50 % calc ( 25 % + 10px ) " but got "50 % top calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [center top 20px] at (0.75) should be [50% calc(15px + 12.5%)] assert_equals: expected "50 % calc ( 12.5 % + 15px ) " but got "50 % top calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [center top 20px] at (1) should be [50% 20px] assert_equals: expected "50 % 20px " but got "50 % top 20px "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center top 20px] at (0.25) should be [50% calc(5px + 37.5%)] assert_equals: expected "50 % calc ( 37.5 % + 5px ) " but got "50 % top calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center top 20px] at (0.5) should be [50% calc(10px + 25%)] assert_equals: expected "50 % calc ( 25 % + 10px ) " but got "50 % top calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center top 20px] at (0.75) should be [50% calc(15px + 12.5%)] assert_equals: expected "50 % calc ( 12.5 % + 15px ) " but got "50 % top calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center top 20px] at (1) should be [50% 20px] assert_equals: expected "50 % 20px " but got "50 % top 20px "
PASS CSS Animations: property <background-position> from [center center] to [center top 20px] at (0) should be [50% 50%] 
FAIL CSS Animations: property <background-position> from [center center] to [center top 20px] at (0.25) should be [50% calc(5px + 37.5%)] assert_equals: expected "50 % calc ( 37.5 % + 5px ) " but got "50 % calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [center top 20px] at (0.5) should be [50% calc(10px + 25%)] assert_equals: expected "50 % calc ( 25 % + 10px ) " but got "50 % calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [center top 20px] at (0.75) should be [50% calc(15px + 12.5%)] assert_equals: expected "50 % calc ( 12.5 % + 15px ) " but got "50 % calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
PASS CSS Animations: property <background-position> from [center center] to [center top 20px] at (1) should be [50% 20px] 
PASS Web Animations: property <background-position> from [center center] to [center top 20px] at (0) should be [50% 50%] 
FAIL Web Animations: property <background-position> from [center center] to [center top 20px] at (0.25) should be [50% calc(5px + 37.5%)] assert_equals: expected "50 % calc ( 37.5 % + 5px ) " but got "50 % calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [center top 20px] at (0.5) should be [50% calc(10px + 25%)] assert_equals: expected "50 % calc ( 25 % + 10px ) " but got "50 % calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [center top 20px] at (0.75) should be [50% calc(15px + 12.5%)] assert_equals: expected "50 % calc ( 12.5 % + 15px ) " but got "50 % calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
PASS Web Animations: property <background-position> from [center center] to [center top 20px] at (1) should be [50% 20px] 
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px top 20px] at (0.25) should be [calc(-5px + 62.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 37.5 % + 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px top 20px] at (0.5) should be [calc(-10px + 75%) calc(10px + 25%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 25 % + 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px top 20px] at (0.75) should be [calc(-15px + 87.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 12.5 % + 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px top 20px] at (1) should be [calc(-20px + 100%) 20px] assert_equals: expected "calc ( 100 % - 20px ) 20px " but got "left calc ( 100 % - 20px ) top 20px "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px top 20px] at (0.25) should be [calc(-5px + 62.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 37.5 % + 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px top 20px] at (0.5) should be [calc(-10px + 75%) calc(10px + 25%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 25 % + 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px top 20px] at (0.75) should be [calc(-15px + 87.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 12.5 % + 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px top 20px] at (1) should be [calc(-20px + 100%) 20px] assert_equals: expected "calc ( 100 % - 20px ) 20px " but got "left calc ( 100 % - 20px ) top 20px "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px top 20px] at (0.25) should be [calc(-5px + 62.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 37.5 % + 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px top 20px] at (0.5) should be [calc(-10px + 75%) calc(10px + 25%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 25 % + 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px top 20px] at (0.75) should be [calc(-15px + 87.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 12.5 % + 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px top 20px] at (1) should be [calc(-20px + 100%) 20px] assert_equals: expected "calc ( 100 % - 20px ) 20px " but got "left calc ( 100 % - 20px ) 20px "
FAIL Web Animations: property <background-position> from [center center] to [right 20px top 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL Web Animations: property <background-position> from [center center] to [right 20px top 20px] at (0.25) should be [calc(-5px + 62.5%) calc(5px + 37.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 37.5 % + 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [right 20px top 20px] at (0.5) should be [calc(-10px + 75%) calc(10px + 25%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 25 % + 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [right 20px top 20px] at (0.75) should be [calc(-15px + 87.5%) calc(15px + 12.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 12.5 % + 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [right 20px top 20px] at (1) should be [calc(-20px + 100%) 20px] assert_equals: expected "calc ( 100 % - 20px ) 20px " but got "left calc ( 100 % - 20px ) 20px "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px center] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px center] at (0.25) should be [calc(5px + 37.5%) 50%] assert_equals: expected "calc ( 37.5 % + 5px ) 50 % " but got "left calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px center] at (0.5) should be [calc(10px + 25%) 50%] assert_equals: expected "calc ( 25 % + 10px ) 50 % " but got "left calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px center] at (0.75) should be [calc(15px + 12.5%) 50%] assert_equals: expected "calc ( 12.5 % + 15px ) 50 % " but got "left calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px center] at (1) should be [20px 50%] assert_equals: expected "20px 50 % " but got "left 20px 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px center] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px center] at (0.25) should be [calc(5px + 37.5%) 50%] assert_equals: expected "calc ( 37.5 % + 5px ) 50 % " but got "left calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px center] at (0.5) should be [calc(10px + 25%) 50%] assert_equals: expected "calc ( 25 % + 10px ) 50 % " but got "left calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px center] at (0.75) should be [calc(15px + 12.5%) 50%] assert_equals: expected "calc ( 12.5 % + 15px ) 50 % " but got "left calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px center] at (1) should be [20px 50%] assert_equals: expected "20px 50 % " but got "left 20px 50 % "
PASS CSS Animations: property <background-position> from [center center] to [left 20px center] at (0) should be [50% 50%] 
FAIL CSS Animations: property <background-position> from [center center] to [left 20px center] at (0.25) should be [calc(5px + 37.5%) 50%] assert_equals: expected "calc ( 37.5 % + 5px ) 50 % " but got "calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px center] at (0.5) should be [calc(10px + 25%) 50%] assert_equals: expected "calc ( 25 % + 10px ) 50 % " but got "calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px center] at (0.75) should be [calc(15px + 12.5%) 50%] assert_equals: expected "calc ( 12.5 % + 15px ) 50 % " but got "calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) 50 % "
PASS CSS Animations: property <background-position> from [center center] to [left 20px center] at (1) should be [20px 50%] 
PASS Web Animations: property <background-position> from [center center] to [left 20px center] at (0) should be [50% 50%] 
FAIL Web Animations: property <background-position> from [center center] to [left 20px center] at (0.25) should be [calc(5px + 37.5%) 50%] assert_equals: expected "calc ( 37.5 % + 5px ) 50 % " but got "calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) 50 % "
FAIL Web Animations: property <background-position> from [center center] to [left 20px center] at (0.5) should be [calc(10px + 25%) 50%] assert_equals: expected "calc ( 25 % + 10px ) 50 % " but got "calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) 50 % "
FAIL Web Animations: property <background-position> from [center center] to [left 20px center] at (0.75) should be [calc(15px + 12.5%) 50%] assert_equals: expected "calc ( 12.5 % + 15px ) 50 % " but got "calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) 50 % "
PASS Web Animations: property <background-position> from [center center] to [left 20px center] at (1) should be [20px 50%] 
PASS CSS Transitions: property <background-position> from [center center] to [center center] at (0) should be [50% 50%] 
PASS CSS Transitions: property <background-position> from [center center] to [center center] at (0.25) should be [50% 50%] 
PASS CSS Transitions: property <background-position> from [center center] to [center center] at (0.5) should be [50% 50%] 
PASS CSS Transitions: property <background-position> from [center center] to [center center] at (0.75) should be [50% 50%] 
PASS CSS Transitions: property <background-position> from [center center] to [center center] at (1) should be [50% 50%] 
PASS CSS Transitions with transition: all: property <background-position> from [center center] to [center center] at (0) should be [50% 50%] 
PASS CSS Transitions with transition: all: property <background-position> from [center center] to [center center] at (0.25) should be [50% 50%] 
PASS CSS Transitions with transition: all: property <background-position> from [center center] to [center center] at (0.5) should be [50% 50%] 
PASS CSS Transitions with transition: all: property <background-position> from [center center] to [center center] at (0.75) should be [50% 50%] 
PASS CSS Transitions with transition: all: property <background-position> from [center center] to [center center] at (1) should be [50% 50%] 
PASS CSS Animations: property <background-position> from [center center] to [center center] at (0) should be [50% 50%] 
PASS CSS Animations: property <background-position> from [center center] to [center center] at (0.25) should be [50% 50%] 
PASS CSS Animations: property <background-position> from [center center] to [center center] at (0.5) should be [50% 50%] 
PASS CSS Animations: property <background-position> from [center center] to [center center] at (0.75) should be [50% 50%] 
PASS CSS Animations: property <background-position> from [center center] to [center center] at (1) should be [50% 50%] 
PASS Web Animations: property <background-position> from [center center] to [center center] at (0) should be [50% 50%] 
PASS Web Animations: property <background-position> from [center center] to [center center] at (0.25) should be [50% 50%] 
PASS Web Animations: property <background-position> from [center center] to [center center] at (0.5) should be [50% 50%] 
PASS Web Animations: property <background-position> from [center center] to [center center] at (0.75) should be [50% 50%] 
PASS Web Animations: property <background-position> from [center center] to [center center] at (1) should be [50% 50%] 
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px center] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px center] at (0.25) should be [calc(-5px + 62.5%) 50%] assert_equals: expected "calc ( 62.5 % - 5px ) 50 % " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px center] at (0.5) should be [calc(-10px + 75%) 50%] assert_equals: expected "calc ( 75 % - 10px ) 50 % " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px center] at (0.75) should be [calc(-15px + 87.5%) 50%] assert_equals: expected "calc ( 87.5 % - 15px ) 50 % " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px center] at (1) should be [calc(-20px + 100%) 50%] assert_equals: expected "calc ( 100 % - 20px ) 50 % " but got "left calc ( 100 % - 20px ) 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px center] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px center] at (0.25) should be [calc(-5px + 62.5%) 50%] assert_equals: expected "calc ( 62.5 % - 5px ) 50 % " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px center] at (0.5) should be [calc(-10px + 75%) 50%] assert_equals: expected "calc ( 75 % - 10px ) 50 % " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px center] at (0.75) should be [calc(-15px + 87.5%) 50%] assert_equals: expected "calc ( 87.5 % - 15px ) 50 % " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px center] at (1) should be [calc(-20px + 100%) 50%] assert_equals: expected "calc ( 100 % - 20px ) 50 % " but got "left calc ( 100 % - 20px ) 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px center] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px center] at (0.25) should be [calc(-5px + 62.5%) 50%] assert_equals: expected "calc ( 62.5 % - 5px ) 50 % " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px center] at (0.5) should be [calc(-10px + 75%) 50%] assert_equals: expected "calc ( 75 % - 10px ) 50 % " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px center] at (0.75) should be [calc(-15px + 87.5%) 50%] assert_equals: expected "calc ( 87.5 % - 15px ) 50 % " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px center] at (1) should be [calc(-20px + 100%) 50%] assert_equals: expected "calc ( 100 % - 20px ) 50 % " but got "left calc ( 100 % - 20px ) 50 % "
FAIL Web Animations: property <background-position> from [center center] to [right 20px center] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % 50 % "
FAIL Web Animations: property <background-position> from [center center] to [right 20px center] at (0.25) should be [calc(-5px + 62.5%) 50%] assert_equals: expected "calc ( 62.5 % - 5px ) 50 % " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) 50 % "
FAIL Web Animations: property <background-position> from [center center] to [right 20px center] at (0.5) should be [calc(-10px + 75%) 50%] assert_equals: expected "calc ( 75 % - 10px ) 50 % " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) 50 % "
FAIL Web Animations: property <background-position> from [center center] to [right 20px center] at (0.75) should be [calc(-15px + 87.5%) 50%] assert_equals: expected "calc ( 87.5 % - 15px ) 50 % " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) 50 % "
FAIL Web Animations: property <background-position> from [center center] to [right 20px center] at (1) should be [calc(-20px + 100%) 50%] assert_equals: expected "calc ( 100 % - 20px ) 50 % " but got "left calc ( 100 % - 20px ) 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px bottom 20px] at (0.25) should be [calc(5px + 37.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 62.5 % - 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px bottom 20px] at (0.5) should be [calc(10px + 25%) calc(-10px + 75%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 75 % - 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px bottom 20px] at (0.75) should be [calc(15px + 12.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 87.5 % - 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [left 20px bottom 20px] at (1) should be [20px calc(-20px + 100%)] assert_equals: expected "20px calc ( 100 % - 20px ) " but got "left 20px top calc ( 100 % - 20px ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px bottom 20px] at (0.25) should be [calc(5px + 37.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 62.5 % - 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px bottom 20px] at (0.5) should be [calc(10px + 25%) calc(-10px + 75%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 75 % - 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px bottom 20px] at (0.75) should be [calc(15px + 12.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 87.5 % - 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [left 20px bottom 20px] at (1) should be [20px calc(-20px + 100%)] assert_equals: expected "20px calc ( 100 % - 20px ) " but got "left 20px top calc ( 100 % - 20px ) "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0.25) should be [calc(5px + 37.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 62.5 % - 5px ) " but got "calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0.5) should be [calc(10px + 25%) calc(-10px + 75%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 75 % - 10px ) " but got "calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0.75) should be [calc(15px + 12.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 87.5 % - 15px ) " but got "calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (1) should be [20px calc(-20px + 100%)] assert_equals: expected "20px calc ( 100 % - 20px ) " but got "20px top calc ( 100 % - 20px ) "
FAIL Web Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL Web Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0.25) should be [calc(5px + 37.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 37.5 % + 5px ) calc ( 62.5 % - 5px ) " but got "calc ( ( 50 % * 0.75 ) + ( 20px * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0.5) should be [calc(10px + 25%) calc(-10px + 75%)] assert_equals: expected "calc ( 25 % + 10px ) calc ( 75 % - 10px ) " but got "calc ( ( 50 % * 0.5 ) + ( 20px * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (0.75) should be [calc(15px + 12.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 12.5 % + 15px ) calc ( 87.5 % - 15px ) " but got "calc ( ( 50 % * 0.25 ) + ( 20px * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [left 20px bottom 20px] at (1) should be [20px calc(-20px + 100%)] assert_equals: expected "20px calc ( 100 % - 20px ) " but got "20px top calc ( 100 % - 20px ) "
FAIL CSS Transitions: property <background-position> from [center center] to [center bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [center bottom 20px] at (0.25) should be [50% calc(-5px + 62.5%)] assert_equals: expected "50 % calc ( 62.5 % - 5px ) " but got "50 % top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [center bottom 20px] at (0.5) should be [50% calc(-10px + 75%)] assert_equals: expected "50 % calc ( 75 % - 10px ) " but got "50 % top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [center bottom 20px] at (0.75) should be [50% calc(-15px + 87.5%)] assert_equals: expected "50 % calc ( 87.5 % - 15px ) " but got "50 % top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [center bottom 20px] at (1) should be [50% calc(-20px + 100%)] assert_equals: expected "50 % calc ( 100 % - 20px ) " but got "50 % top calc ( 100 % - 20px ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center bottom 20px] at (0.25) should be [50% calc(-5px + 62.5%)] assert_equals: expected "50 % calc ( 62.5 % - 5px ) " but got "50 % top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center bottom 20px] at (0.5) should be [50% calc(-10px + 75%)] assert_equals: expected "50 % calc ( 75 % - 10px ) " but got "50 % top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center bottom 20px] at (0.75) should be [50% calc(-15px + 87.5%)] assert_equals: expected "50 % calc ( 87.5 % - 15px ) " but got "50 % top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [center bottom 20px] at (1) should be [50% calc(-20px + 100%)] assert_equals: expected "50 % calc ( 100 % - 20px ) " but got "50 % top calc ( 100 % - 20px ) "
FAIL CSS Animations: property <background-position> from [center center] to [center bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [center bottom 20px] at (0.25) should be [50% calc(-5px + 62.5%)] assert_equals: expected "50 % calc ( 62.5 % - 5px ) " but got "50 % top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [center bottom 20px] at (0.5) should be [50% calc(-10px + 75%)] assert_equals: expected "50 % calc ( 75 % - 10px ) " but got "50 % top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [center bottom 20px] at (0.75) should be [50% calc(-15px + 87.5%)] assert_equals: expected "50 % calc ( 87.5 % - 15px ) " but got "50 % top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [center bottom 20px] at (1) should be [50% calc(-20px + 100%)] assert_equals: expected "50 % calc ( 100 % - 20px ) " but got "50 % top calc ( 100 % - 20px ) "
FAIL Web Animations: property <background-position> from [center center] to [center bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "50 % top 50 % "
FAIL Web Animations: property <background-position> from [center center] to [center bottom 20px] at (0.25) should be [50% calc(-5px + 62.5%)] assert_equals: expected "50 % calc ( 62.5 % - 5px ) " but got "50 % top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [center bottom 20px] at (0.5) should be [50% calc(-10px + 75%)] assert_equals: expected "50 % calc ( 75 % - 10px ) " but got "50 % top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [center bottom 20px] at (0.75) should be [50% calc(-15px + 87.5%)] assert_equals: expected "50 % calc ( 87.5 % - 15px ) " but got "50 % top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [center bottom 20px] at (1) should be [50% calc(-20px + 100%)] assert_equals: expected "50 % calc ( 100 % - 20px ) " but got "50 % top calc ( 100 % - 20px ) "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px bottom 20px] at (0.25) should be [calc(-5px + 62.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 62.5 % - 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px bottom 20px] at (0.5) should be [calc(-10px + 75%) calc(-10px + 75%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 75 % - 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px bottom 20px] at (0.75) should be [calc(-15px + 87.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 87.5 % - 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Transitions: property <background-position> from [center center] to [right 20px bottom 20px] at (1) should be [calc(-20px + 100%) calc(-20px + 100%)] assert_equals: expected "calc ( 100 % - 20px ) calc ( 100 % - 20px ) " but got "left calc ( 100 % - 20px ) top calc ( 100 % - 20px ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px bottom 20px] at (0.25) should be [calc(-5px + 62.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 62.5 % - 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px bottom 20px] at (0.5) should be [calc(-10px + 75%) calc(-10px + 75%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 75 % - 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px bottom 20px] at (0.75) should be [calc(-15px + 87.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 87.5 % - 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Transitions with transition: all: property <background-position> from [center center] to [right 20px bottom 20px] at (1) should be [calc(-20px + 100%) calc(-20px + 100%)] assert_equals: expected "calc ( 100 % - 20px ) calc ( 100 % - 20px ) " but got "left calc ( 100 % - 20px ) top calc ( 100 % - 20px ) "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0.25) should be [calc(-5px + 62.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 62.5 % - 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0.5) should be [calc(-10px + 75%) calc(-10px + 75%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 75 % - 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0.75) should be [calc(-15px + 87.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 87.5 % - 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL CSS Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (1) should be [calc(-20px + 100%) calc(-20px + 100%)] assert_equals: expected "calc ( 100 % - 20px ) calc ( 100 % - 20px ) " but got "left calc ( 100 % - 20px ) top calc ( 100 % - 20px ) "
FAIL Web Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0) should be [50% 50%] assert_equals: expected "50 % 50 % " but got "left 50 % top 50 % "
FAIL Web Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0.25) should be [calc(-5px + 62.5%) calc(-5px + 62.5%)] assert_equals: expected "calc ( 62.5 % - 5px ) calc ( 62.5 % - 5px ) " but got "left calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) top calc ( ( 50 % * 0.75 ) + ( ( 100 % - 20px ) * 0.25 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0.5) should be [calc(-10px + 75%) calc(-10px + 75%)] assert_equals: expected "calc ( 75 % - 10px ) calc ( 75 % - 10px ) " but got "left calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) top calc ( ( 50 % * 0.5 ) + ( ( 100 % - 20px ) * 0.5 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (0.75) should be [calc(-15px + 87.5%) calc(-15px + 87.5%)] assert_equals: expected "calc ( 87.5 % - 15px ) calc ( 87.5 % - 15px ) " but got "left calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) top calc ( ( 50 % * 0.25 ) + ( ( 100 % - 20px ) * 0.75 ) ) "
FAIL Web Animations: property <background-position> from [center center] to [right 20px bottom 20px] at (1) should be [calc(-20px + 100%) calc(-20px + 100%)] assert_equals: expected "calc ( 100 % - 20px ) calc ( 100 % - 20px ) " but got "left calc ( 100 % - 20px ) top calc ( 100 % - 20px ) "
PASS CSS Transitions: property <background-position> from [center] to [bottom] at (0) should be [50% 50%] 
PASS CSS Transitions: property <background-position> from [center] to [bottom] at (0.25) should be [50% 62.5%] 
PASS CSS Transitions: property <background-position> from [center] to [bottom] at (0.5) should be [50% 75%] 
PASS CSS Transitions: property <background-position> from [center] to [bottom] at (0.75) should be [50% 87.5%] 
PASS CSS Transitions: property <background-position> from [center] to [bottom] at (1) should be [50% 100%] 
PASS CSS Transitions with transition: all: property <background-position> from [center] to [bottom] at (0) should be [50% 50%] 
PASS CSS Transitions with transition: all: property <background-position> from [center] to [bottom] at (0.25) should be [50% 62.5%] 
PASS CSS Transitions with transition: all: property <background-position> from [center] to [bottom] at (0.5) should be [50% 75%] 
PASS CSS Transitions with transition: all: property <background-position> from [center] to [bottom] at (0.75) should be [50% 87.5%] 
PASS CSS Transitions with transition: all: property <background-position> from [center] to [bottom] at (1) should be [50% 100%] 
PASS CSS Animations: property <background-position> from [center] to [bottom] at (0) should be [50% 50%] 
PASS CSS Animations: property <background-position> from [center] to [bottom] at (0.25) should be [50% 62.5%] 
PASS CSS Animations: property <background-position> from [center] to [bottom] at (0.5) should be [50% 75%] 
PASS CSS Animations: property <background-position> from [center] to [bottom] at (0.75) should be [50% 87.5%] 
PASS CSS Animations: property <background-position> from [center] to [bottom] at (1) should be [50% 100%] 
PASS Web Animations: property <background-position> from [center] to [bottom] at (0) should be [50% 50%] 
PASS Web Animations: property <background-position> from [center] to [bottom] at (0.25) should be [50% 62.5%] 
PASS Web Animations: property <background-position> from [center] to [bottom] at (0.5) should be [50% 75%] 
PASS Web Animations: property <background-position> from [center] to [bottom] at (0.75) should be [50% 87.5%] 
PASS Web Animations: property <background-position> from [center] to [bottom] at (1) should be [50% 100%] 
